+++ examples/workspace/target/debug/libmember1.d
@@ -1,0 +1,1 @@
+/home/pi/workspace/projects/rust/cargo-make/examples/workspace/target/debug/libmember1.rlib: /home/pi/workspace/projects/rust/cargo-make/examples/workspace/member1/src/lib.rs