+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.32.8
 
+* Load cargo aliases as tasks #484
 * Support script content as a simple string instead of an array.
 * Support script content as fielded sections (pre/main/post) to enable sharing of common script content between tasks #473
 * The load_script and install_script attribute types supports all script input types (string/vec/file/sections).